// packages/game-logic/src/types.ts
export enum Suit {
  Hearts = 'Hearts',
  Diamonds = 'Diamonds',
  Clubs = 'Clubs',
  Spades = 'Spades',
}

export enum Rank {
  Two = '2',
  Three = '3',
  Four = '4',
  Five = '5',
  Six = '6',
  Seven = '7',
  Eight = '8',
  Nine = '9',
  Ten = '10',
  Jack = 'J',
  Queen = 'Q',
  King = 'K',
  Ace = 'A',
}

export interface Card {
  suit: Suit;
  rank: Rank;
}

export enum PokerHand {
    HighCard = 'High Card',
    Pair = 'Pair',
    TwoPair = 'Two Pair',
    ThreeOfAKind = 'Three of a Kind',
    Straight = 'Straight',
    Flush = 'Flush',
    FullHouse = 'Full House',
    FourOfAKind = 'Four of a Kind',
    StraightFlush = 'Straight Flush',
    RoyalFlush = 'Royal Flush',
}